Implemented max_pooling in a conv_block
